new files distribution, bug fixes
